1. changes in performance report 2. search by serial no 3. status dropdown instead of assignee dropdown 4. add new contact changes in register ticket form 5. show all tickets changes that can only seen by admin and superadmin 6. fromdate todate filter applied on dashboard

+++ lib/viewmodel/workPerformance_viewmodel.dart
@@ -77,8 +77,8 @@ class WorkPerformanceViewModel extends ChangeNotifier {
   int get overdue =>
       summary?.overdue ?? 0;
 
-  int get avgResolutionTime =>
-      summary?.avgResolutionTime ?? 0;
+  double get avgResolutionTime =>
+      summary?.avgResolutionTime ?? 0.0;
 
   int get productivityScore =>
       summary?.productivityScore ?? 0;
